--- src/main/java/Jeffybezos.java
@@ -1,10 +1,0 @@
-public class Jeffybezos {
-    public static void main(String[] args) {
-        System.out.println("________________________________\n");
-        System.out.println("Hello! I'm jeffybezos\n"
-                + "What can I do for you?\n");
-        System.out.println("________________________________\n");
-        System.out.println("Bye. Hope to see you again soon!\n");
-        System.out.println("________________________________\n");
-    }
-}
